factor out env var logic
