+++ crates/rv/src/config.rs
@@ -1,3 +1,8 @@
+use std::{
+    env::{JoinPathsError, join_paths, split_paths},
+    path::PathBuf,
+};
+
 use camino::{Utf8Path, Utf8PathBuf};
 use tracing::{debug, instrument};
 
@@ -18,6 +23,10 @@ pub enum Error {
     IoError(#[from] std::io::Error),
     #[error(transparent)]
     RequestError(#[from] RequestError),
+    #[error(transparent)]
+    EnvError(#[from] std::env::VarError),
+    #[error(transparent)]
+    JoinPathsError(#[from] JoinPathsError),
 }
 
 type Result<T> = miette::Result<T, Error>;
@@ -114,4 +123,72 @@ pub fn find_project_dir(current_dir: Utf8PathBuf, root: Utf8PathBuf) -> Option<U
             return None;
         }
     }
+}
+
+const ENV_VARS: [&str; 7] = [
+    "RUBY_ROOT",
+    "RUBY_ENGINE",
+    "RUBY_VERSION",
+    "RUBYOPT",
+    "GEM_ROOT",
+    "GEM_HOME",
+    "GEM_PATH",
+];
+
+#[allow(clippy::type_complexity)]
+pub fn env_for(ruby: &Option<Ruby>) -> Result<(Vec<&'static str>, Vec<(&'static str, String)>)> {
+    let mut unset: Vec<_> = ENV_VARS.into();
+    let mut set: Vec<(&'static str, String)> = vec![];
+
+    let mut insert = |var: &'static str, val: String| {
+        // PATH is never in the list to unset
+        if let Some(i) = unset.iter().position(|i| *i == var) {
+            unset.remove(i);
+        }
+
+        set.push((var, val));
+    };
+
+    let mut paths = std::env::var("PATH").map(|p| split_paths(&p).collect::<Vec<_>>())?;
+
+    let old_ruby_paths: Vec<PathBuf> = ["RUBY_ROOT", "GEM_ROOT", "GEM_HOME"]
+        .iter()
+        .filter_map(|var| std::env::var(var).ok())
+        .map(|p| std::path::Path::new(&p).join("bin"))
+        .collect();
+
+    let old_gem_paths: Vec<PathBuf> =
+        std::env::var("GEM_PATH").map_or_else(|_| vec![], |p| split_paths(&p).collect::<Vec<_>>());
+
+    // Remove old Ruby and Gem paths from PATH
+    paths.retain(|p| !old_ruby_paths.contains(p) && !old_gem_paths.contains(p));
+
+    if let Some(ruby) = ruby {
+        let mut gem_paths = vec![];
+        paths.insert(0, ruby.bin_path().into());
+        insert("RUBY_ROOT", ruby.path.to_string());
+        insert("RUBY_ENGINE", ruby.version.engine.name().into());
+        insert("RUBY_VERSION", ruby.version.number());
+        if let Some(gem_home) = ruby.gem_home() {
+            paths.insert(0, gem_home.join("bin").into());
+            gem_paths.insert(0, gem_home.join("bin"));
+            insert("GEM_HOME", gem_home.into_string());
+        }
+        if let Some(gem_root) = ruby.gem_root() {
+            paths.insert(0, gem_root.join("bin").into());
+            gem_paths.insert(0, gem_root.join("bin"));
+            insert("GEM_ROOT", gem_root.into_string());
+        }
+        let gem_path = join_paths(gem_paths)?;
+        if let Some(gem_path) = gem_path.to_str() {
+            insert("GEM_PATH", gem_path.into());
+        }
+    }
+
+    let path = join_paths(paths)?;
+    if let Some(path) = path.to_str() {
+        insert("PATH", path.into());
+    }
+
+    Ok((unset, set))
 }
